requirements.txt
marshmallow~=3.16.0
requests==2.27.1
robotframework~=4.1.3
robotframework-assertion-engine==0.5.1
#robotframework-browser==12.4.0
robotframework-pythonlibcore==3.0.0
robotframework-requests==0.9.3
selenium~=4.6.0
rpaframework==19.1.1
webdriver-manager==3.8.4
